+++ app/src/main/java/com/verbanode/mobile/network/ConnectionState.kt
@@ -1,8 +1,0 @@
-package com.verbanode.mobile.network
-
-enum class ConnectionState {
-    DISCONNECTED,
-    CONNECTING,
-    CONNECTED,
-    RECONNECTING,
-}